solved all issues mentioned in the PR.
